refactored AlicaTimer
